pyproject.toml
[tool.poetry]
name = "patent-search-agent"
version = "0.1.0"
description = "Multi-agent system for patent prior art search using LangGraph"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
langgraph = "^0.2.0"
langchain-anthropic = "^0.2.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
tavily-python = "^0.5.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
pydantic = "^2.8.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
